requirements.txt
Scrapy==0.24.2
dateutils==0.6.6
pdfminer
